suppress implicit cast std::size_t -> int warning

+++ src/koatc/logger/wchar_wrapper.cxx
@@ -3,10 +3,18 @@
 
 namespace turenar::koatc::logger::internal {
 std::string convert_to_utf8(const std::wstring_view& s) {
-	int length = ::WideCharToMultiByte(CP_UTF8, 0, s.data(), s.size(), nullptr, 0, nullptr, nullptr);
+	int length = ::WideCharToMultiByte(CP_UTF8, 0, s.data(), static_cast<int>(s.size()), nullptr, 0, nullptr, nullptr);
 	std::string ret;
 	ret.resize(length);
-	length = ::WideCharToMultiByte(CP_UTF8, 0, s.data(), s.size(), ret.data(), ret.size(), nullptr, nullptr);
+	length = ::WideCharToMultiByte(
+			CP_UTF8,
+			0,
+			s.data(),
+			static_cast<int>(s.size()),
+			ret.data(),
+			static_cast<int>(ret.size()),
+			nullptr,
+			nullptr);
 	ret.resize(length);
 	return ret;
 }
